PatrollingGuard now uses pathfinding for chasing intruders, small changes to end of game mechanics
