+++ simulation/agent.py
@@ -1,7 +1,6 @@
 from typing import NewType, List, Tuple
 from abc import ABCMeta, abstractmethod
 import math
-import random
 import vectormath as vmath
 
 from .util import Position
@@ -45,6 +44,7 @@ class Agent(metaclass=ABCMeta):
         self.view_range: float = 6.0
         self.view_angle: float = 45.0
         self.turn_speed: float = 180
+        
         # to keep track of movement commands and execute them in the background
         self._move_target: float = 0
         self._turn_target: float = 0
@@ -229,42 +229,25 @@ class Agent(metaclass=ABCMeta):
         """ Agent logic goes here """
         pass
 
-
 # TODO: implement sentry tower
 class GuardAgent(Agent):
     def __init__(self) -> None:
         super().__init__()
-        self.color = (0.0, 1.0, 0.0)
+        self.color = (0, 1, 0) # green
         self.view_range: float = 6.0
         
-    def make_patrol_route(self) -> List['Position']:
-        width  = [1.5, self.map.width - 1.5]
-        height = [1.5, self.map.height - 1.5]
-        corner_points = [Position(vmath.Vector2(a, b)) for a in width for b in height]
-        
-        patrol_route = []
-        
-        for i in range(3):
-            route_point = corner_points.pop(random.randint(0,len(corner_points)-1))
-            patrol_route.append(route_point)
-        return patrol_route
-
     def setup(self, world):
         super().setup(world)
         self.other_guards = [vision.AgentView(guard) for ID, guard in self._world.guards.items() if not ID == self.ID]
-        self.patrol_route = self.make_patrol_route()
-        self.target_idx = 0
-        self.target = self.patrol_route[self.target_idx]
-        print('Guard', self.ID, 'Patrolling Route:', self.patrol_route)
         
 # TODO: implement sprinting
 class IntruderAgent(Agent):
     def __init__(self) -> None:
         super().__init__()
-        self.color = (1.0, 0.0, 0.0) # needed?
+        self.color = (1, 1, 0) # yellow
         self.view_range: float = 7.5
         self.target = Position(vmath.Vector2((1.5, 1.5))) # must be .5 (center of tile)
-        
+                
         # are we captured yet?
         self.is_captured = False
         self._prev_is_captured = False
@@ -285,24 +268,4 @@ class IntruderAgent(Agent):
     def on_reached_target(self) -> None:
         """ Called once the agent has reached its target """
         pass
-
-    def tick(self, seen_agents, noises):
-        if self.is_captured:
-            # make sure we only run the `on_captured` handler once
-            if not self._prev_is_captured:
-                self.on_captured()
-                self._prev_is_captured = True
-
-            # don't run any other agent code if we're captured
-            return
-        elif self.reached_target:
-            # make sure we only run the `on_captured` handler once
-            if not self._prev_reached_target:
-                self.on_reached_target()
-                self._prev_reached_target = True
-            
-            # don't run any other code - the intruder won
-            return 
-        else:
-            # if we're not captured or have reached the target then just proceed as usual
-            super().tick(seen_agents, noises)
+
